+++ tests/integ/loop_ai_test.sh
@@ -4,6 +4,6 @@ set -e
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 pushd ${SCRIPT_DIR} > /dev/null
 for i in {1..100};do
-  ./ai_test.sh
+  ./ai_test.sh ${i}
 done
 popd